requirements.txt
Django==3.1.4
django-allauth==0.44
djangorestframework==3.12.2
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
